pyproject.toml
[project]
name = "clive-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fire>=0.7.0",
    "httpx>=0.28.1",
    "litellm>=1.65.4.post1",
    "python-dotenv>=1.1.0",
    "smolagents>=1.13.0",
]

[dependency-groups]
dev = [
    "autopep8>=2.3.2",
    "pytest>=8.3.5",
]
